Decouple `WantsOutputs` v2 and v1 receiver states
Introduces `core::receive::WantsOutputs` as a common abstraction to
decouple v1 and v2 typestates.
